# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: File Server API
  version: 1.0.0
  description: 'Generated from the provided Spring Boot controllers.

    Note: Controller paths use a property ''${api.base-path}''. In this spec,

    the server URL includes a variable ''{basePath}'' (default ''/api/v1'').

    '
servers:
- url: http://localhost:8080{basePath}
  description: Local development
  variables:
    basePath:
      default: /api/v1
      description: Matches your Spring `${api.base-path}` property
...
paths:
...
  /maintenance-job/start:
    post:
      tags:
      - maintenance-job
      summary: Start maintenance job
      operationId: startMaintenanceJob
      responses:
        '200':
          description: Job started
          content:
            text/plain:
              schema:
                type: string
        '409':
          description: Job already running
          content:
            text/plain:
              schema:
                type: string
  /maintenance-job/stop:
    post:
      tags:
      - maintenance-job
      summary: Stop maintenance job
      operationId: stopMaintenanceJob
      responses:
        '200':
          description: Job stopped
          content:
            text/plain:
              schema:
                type: string
        '409':
          description: No job running
          content:
            text/plain:
              schema:
                type: string
...